5th Commit - In-Memory-App

+++ src/main/java/io/codelex/flightplanner/RestServices/Controllers/CustomerFlightController.java
@@ -31,7 +31,7 @@ public class CustomerFlightController {
     }
 
     @GetMapping("flights/{id}")
-    public Flight findFlightById(@PathVariable int id) {
+    public Flight findFlightById(@PathVariable long id) {
         return customerFlightService.findFlightById(id);
     }
 }
