+++ src/main/java/io/codelex/flightplanner/RestServices/Services/CustomerFlightService.java
@@ -21,7 +21,7 @@ public class CustomerFlightService {
         return flightRepository.searchAirports(search);
     }
 
-    public Flight findFlightById(int id) {
+    public Flight findFlightById(long id) {
         return flightRepository.fetchFlight(id);
     }
